+++ fabric/src/main/java/net/SpectrumFATM/black_archive/fabric/entity/custom/DalekPuppetEntity.java
@@ -1,7 +1,6 @@
 package net.SpectrumFATM.black_archive.fabric.entity.custom;
 
 import net.SpectrumFATM.black_archive.fabric.BlackArchive;
-import net.SpectrumFATM.black_archive.fabric.entity.ModEntities;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.ai.goal.*;
@@ -9,18 +8,18 @@ import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.mob.HostileEntity;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Identifier;
 import net.minecraft.world.World;
 
 import java.util.Random;
+import java.util.UUID;
 
 public class DalekPuppetEntity extends HostileEntity {
-
-    int skinId;
+    private UUID playerUUID;
+    private int skinId;
 
     public DalekPuppetEntity(EntityType<? extends HostileEntity> entityType, World world) {
         super(entityType, world);
-        BlackArchive.LOGGER.info("DalekSlaveEntity initialized");
+        BlackArchive.LOGGER.info("DalekPuppetEntity initialized");
 
         Random random = new Random();
         skinId = random.nextInt(8);
@@ -35,7 +34,7 @@ public class DalekPuppetEntity extends HostileEntity {
         this.goalSelector.add(3, new LookAroundGoal(this));
         this.goalSelector.add(4, new GoToWalkTargetGoal(this, 2.0D));
         this.targetSelector.add(2, new ActiveTargetGoal<>(this, LivingEntity.class, 10, true, false,
-                livingEntity -> !(livingEntity instanceof DalekEntity || livingEntity instanceof DalekPuppetEntity)));
+                livingEntity -> !(livingEntity instanceof DalekEntity || DalekPuppetEntity.class.isAssignableFrom(livingEntity.getClass()))));
     }
 
     public static DefaultAttributeContainer.Builder createDalekSlaveAttributes() {
@@ -54,26 +53,15 @@ public class DalekPuppetEntity extends HostileEntity {
         return false;
     }
 
-    public Identifier getSkin() {
-        switch (this.skinId) {
-            case 0:
-                return new Identifier("minecraft", "textures/entity/player/wide/alex.png");
-            case 1:
-                return new Identifier("minecraft", "textures/entity/player/wide/ari.png");
-            case 2:
-                return new Identifier("minecraft", "textures/entity/player/wide/efe.png");
-            case 3:
-                return new Identifier("minecraft", "textures/entity/player/wide/kai.png");
-            case 4:
-                return new Identifier("minecraft", "textures/entity/player/wide/makena.png");
-            case 5:
-                return new Identifier("minecraft", "textures/entity/player/wide/noor.png");
-            case 6:
-                return new Identifier("minecraft", "textures/entity/player/wide/sunny.png");
-            case 7:
-                return new Identifier("minecraft", "textures/entity/player/wide/zuri.png");
-            default:
-                return new Identifier("minecraft", "textures/entity/player/wide/steve.png");
-        }
+    public void setPlayerUUID(UUID playerUUID) {
+        this.playerUUID = playerUUID;
+    }
+
+    public UUID getPlayerUUID() {
+        return this.playerUUID;
+    }
+
+    public int getSkinId() {
+        return this.skinId;
     }
 }