+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Sonic Screwdrivers now power gravity and oxygen generators.
 - 10th Doctor's Sonic Screwdriver.
 - Steel. Including raw, ingot and blocks.
+- Dalek Puppets now spawn with the players skin when player dies as a result of the nanocloud. This is configurable in client config to prevent excess storage use.
+- Client config added.
 
 ### Fixed
 - Fixed TARDIS gravity generator not working.